Changes the type for PING id from uint32 to uint64.
Required for HTTP2.

This lands server change 60786018 by birenroy.

Minor type updates to QUIC/SpdySession & friends were also required.

Review URL: https://codereview.chromium.org/154353003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@250767 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: bd254d51c03c91f8a42113766d071834be3f8d08

+++ net/spdy/spdy_protocol.h
@@ -365,7 +365,7 @@ typedef uint8 SpdyPriority;
 
 typedef std::map<std::string, std::string> SpdyNameValueBlock;
 
-typedef uint32 SpdyPingId;
+typedef uint64 SpdyPingId;
 
 class SpdyFrame;
 typedef SpdyFrame SpdySerializedFrame;
